Fix/chatroom data fetching (#70)
* fix: fetchNextPage에 async/await 추가

* feat: isLoadingRef 추가

* fix: state를 ref로 바꿔서 버그 해결

* feat: useReversedInfiniteScroll 훅 분리

* rename: hooks/feature를 hooks/queries로 면경
